requirements.txt
Pillow==6.2.0

# Django
django>=3.0
django-extensions==2.2.9

# Json Web Token
pyjwt>=1.7.1

# Django REST Framework
djangorestframework==3.11.0